requirements.txt
numpy>=1.26,<2.0
pandas>=2.2,<3.0
scikit-learn>=1.5,<2.0
xgboost>=2.0,<3.0
lightgbm>=4.3,<5.0
tensorflow>=2.16,<3.0 ; platform_system != "Windows"
torch>=2.3,<3.0
cleanlab>=2.6,<3.0
pyyaml>=6.0,<7.0
pytest>=8.2,<9.0
httpx>=0.27,<0.28
